% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addgroupedseries.R
\name{addgroupedseries}
\alias{addgroupedseries}
\title{Add grouped series.}
\usage{
addgroupedseries(options, data, groupcol, xcol, ycol)
}
\arguments{
\item{options}{Highcharts options for the chart. Includes data, chart type, etc.}

\item{data}{data.frame-like object containing the underlying data. It must already be grouped and summarized.}

\item{groupcol}{Name of the column that will be used to split the data into series. These groups will show in the legend.}

\item{xcol}{Name of the column to be used as the X value.}

\item{ycol}{Name of the column to be used as the Y value.}
}
\value{
options list with series' added.
}
\description{
Seperating data into series is a very common operation. This function takes your grouped data and adds the series'.
}
