% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter_replace.R
\name{hc_add_series_list}
\alias{hc_add_series_list}
\title{Highchart Replacement: hc_add_series_list}
\usage{
hc_add_series_list(options, listofseries)
}
\arguments{
\item{options}{Highcharts options for the chart. Includes data, chart type, etc.}

\item{listofseries}{List of data for one or more series'.}
}
\value{
List that can be passed to other functions.
}
\description{
Added to make code written with Highcharter work with hcslim.
}
