% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_html.R
\name{hc_addgroupedseries}
\alias{hc_addgroupedseries}
\title{hc_addgroupedseries}
\usage{
hc_addgroupedseries(options, data, groupcol, xcol, ycol, zcol = NULL)
}
\arguments{
\item{options}{Your starting Highcharts options.}

\item{data}{Data to pull from.}

\item{groupcol}{Character indicating the column to use for grouping.}

\item{xcol}{Character indicating the column to use for the x value.}

\item{ycol}{Character indicating the column to use for the y value.}

\item{zcol}{Character indicating the column to use for the z value (used in Bubble charts, etc).}
}
\value{
Modified Highcharts options with series' added.
}
\description{
Helper to add a grouped series.
See https://github.com/superchordate/hcslim for more detailed documentation and examples.
}
\details{
Author: Bryce Chamberlain
}
