% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter_replace.R
\name{hc_caption}
\alias{hc_caption}
\title{Highchart Replacement: hc_caption}
\usage{
hc_caption(options, ...)
}
\arguments{
\item{options}{Highcharts options for the chart. Includes data, chart type, etc.}

\item{...}{Option key/value pairs passed to hc_addoption.}
}
\value{
List that can be passed to other functions.
}
\description{
Added to make code written with Highcharter work with hcslim.
}
