% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_html.R
\name{hc_enabletooltips}
\alias{hc_enabletooltips}
\title{hc_enabletooltips}
\usage{
hc_enabletooltips(options)
}
\arguments{
\item{options}{Your starting Highcharts options.}
}
\value{
Modified Highcharts options with tooltips enabled (series > enableMouseTracking > TRUE)
}
\description{
Helper to enable tooltips on a chart. Enabling the tooltip is not intuitive because of how I've set up the defaults, so we have this function to enable them more easily.
I like to use enableMouseTracking = FALSE to remove distractions.
See https://github.com/superchordate/hcslim for more detailed documentation and examples.
}
\details{
Author: Bryce Chamberlain
}
