% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_html.R
\name{hc_html}
\alias{hc_html}
\title{hc_html}
\usage{
hc_html(
  id,
  options,
  class = c("chart", "mapChart", "stockChart", "ganttChart"),
  loadmapfromurl = NULL,
  printjs = FALSE,
  pretty = printjs,
  for_widget = FALSE
)
}
\arguments{
\item{id}{HTML id of the div to put the chart in.}

\item{options}{Options list using standard Highcharts options and hc_markjs to mark Javascript code.}

\item{class}{Type of chart.}

\item{loadmapfromurl}{URL to load your base map from.}

\item{printjs}{Print the generated Javascript. Useful for debugging.}

\item{pretty}{Output readable Javascript.}

\item{for_widget}{Special option for buiding an htmlwidget. Please don't change it.}
}
\value{
HTML script tag with the Javascript to create your Highcharts chart.
}
\description{
Main hcslim function. Converts an R list to Highcharts options and return Highcharts JS.
See https://github.com/superchordate/hcslim for more detailed documentation and examples.
}
\details{
Author: Bryce Chamberlain
}
