% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_html.R
\name{hc_markjs}
\alias{hc_markjs}
\title{hc_markjs}
\usage{
hc_markjs(string)
}
\arguments{
\item{string}{Javascript code as a character.}
}
\value{
Javacript code inside "JS!{string}!" which hc_html will know what to do with.
}
\description{
Marks Javascript code so hc_html knows how to handle it.
See https://github.com/superchordate/hcslim for more detailed documentation and examples.
}
\details{
Author: Bryce Chamberlain
}
