% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter_replace.R
\name{hc_responsive}
\alias{hc_responsive}
\title{Highchart Replacement: hc_responsive}
\usage{
hc_responsive(options, ...)
}
\value{
List that can be passed to other functions.
}
\description{
Added to make code written with Highcharter work with hcslim.
}
