% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_use.R
\name{hc_use}
\alias{hc_use}
\title{hc_use}
\usage{
hc_use(
  maps = FALSE,
  hc_paths = c(ifelse(maps, "maps/highmaps", "highcharts"), "modules/accessibility",
    "highcharts-more", "modules/exporting")
)
}
\arguments{
\item{maps}{Indicate if maps are needed.}

\item{hc_paths}{Relative paths indicating which files to gather. Usually you can ignore this, the defaults should work. You may want to only send "highcharts" if you are running offline (Quarto, RMarkdown, RStudio).}
}
\value{
List of HTML script import snippets to gather files from the CDN.
}
\description{
Helper to import Highcharts files from a CDN.
See https://github.com/superchordate/hcslim for more detailed documentation and examples.
}
\details{
Author: Bryce Chamberlain
}
