% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_widget.R
\name{hc_widget}
\alias{hc_widget}
\title{hcslim Widget (Experimental)}
\usage{
hc_widget(options, width = NULL, height = NULL, elementId = NULL, ...)
}
\arguments{
\item{options}{Options list using standard Highcharts options and hc_markjs to mark Javascript code.}

\item{width}{Standard argument passed to htmlwidgets::createWidget.}

\item{height}{Standard argument passed to htmlwidgets::createWidget.}

\item{elementId}{Standard argument passed to htmlwidgets::createWidget.}

\item{...}{Remaining arguments are passed to hc_html.}
}
\value{
htmlwidgets widget representing your Highcharts chart.
}
\description{
htmlwidget with embedded JS for easier use in Quarto, RMarkdown, RStudio, and R Shiny.
See https://github.com/superchordate/hcslim for more detailed documentation and examples.
}
\details{
Author: Bryce Chamberlain
}
