% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_highcharter.R
\name{hcslim-highcharter}
\alias{hcslim-highcharter}
\alias{highchart}
\alias{hc_xAxis}
\alias{hc_yAxis}
\alias{hc_subtitle}
\alias{hc_tooltip}
\alias{hc_annotations}
\alias{hc_caption}
\alias{hc_chart}
\alias{hc_colorAxis}
\alias{hc_colors}
\alias{hc_credits}
\alias{hc_drilldown}
\alias{hc_exporting}
\alias{hc_labels}
\alias{hc_legend}
\alias{hc_mapNavigation}
\alias{hc_motion}
\alias{hc_navigator}
\alias{hc_pane}
\alias{hc_plotOptions}
\alias{hc_rangeSelector}
\alias{hc_responsive}
\alias{hc_scrollbar}
\alias{hc_size}
\alias{hc_add_series}
\alias{hc_add_series_list}
\alias{renderHighchart}
\alias{highchartOutput}
\alias{list_parse}
\alias{list_parse2}
\alias{hc_addoption}
\alias{hc_addseries}
\title{Highcharter override functions.}
\usage{
highchart()

hc_xAxis(options, ...)

hc_yAxis(options, ...)

hc_subtitle(options, ...)

hc_tooltip(options, ...)

hc_annotations(options, ...)

hc_caption(options, ...)

hc_chart(options, ...)

hc_colorAxis(options, ...)

hc_colors(options, ...)

hc_credits(options, ...)

hc_drilldown(options, ...)

hc_exporting(options, ...)

hc_labels(options, ...)

hc_legend(options, ...)

hc_mapNavigation(options, ...)

hc_motion(options, ...)

hc_navigator(options, ...)

hc_pane(options, ...)

hc_plotOptions(options, ...)

hc_rangeSelector(options, ...)

hc_responsive(options, ...)

hc_scrollbar(options, ...)

hc_size(options, ...)

hc_subtitle(options, ...)

hc_add_series(options, ...)

hc_add_series_list(options, listofseries)

renderHighchart(id, options, class = "chart", prettyjs = FALSE)

highchartOutput(id, height = NULL, width = NULL)

list_parse(x)

list_parse2(x)

hc_addoption(options, name, ...)

hc_addseries(options, series)
}
\arguments{
\item{options}{Options list using standard Highcharts options and hc_markjs to mark Javascript code.}
}
\description{
These functions help with swapping highcharter to hcslim in an existing application.
See https://github.com/superchordate/hcslim for more detailed documentation and examples.
}
\details{
Author: Bryce Chamberlain
}
