% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcslim_view.R
\name{hcslim_view}
\alias{hcslim_view}
\title{View hcslim result in viewer.}
\usage{
hcslim_view(
  options,
  class = c("chart", "mapChart", "stockChart", "ganttChart"),
  usecode_include = ifelse(class[1] == "mapChart", "maps/highmaps", "highcharts"),
  loadmapfromurl = NULL,
  pretty = printjs,
  printjs = FALSE,
  printtempfile = FALSE,
  printhtml = FALSE
)
}
\arguments{
\item{options}{Highcharts options for the chart. Includes data, chart type, etc.}

\item{class}{"chart" in Highcharts.chart(... For Highmaps, it should be "mapChart".}

\item{usecode_include}{Required Highcharts modules that you'd pass to hcslim::usecode().}

\item{loadmapfromurl}{If you are using a map, this is the URL of the topo.json file. Example: 'https://code.highcharts.com/mapdata/countries/us/us-ca-all.topo.json'}

\item{pretty}{Use pretty formatting when converting options to JSON.}

\item{printjs}{Print completed JS to console for troubleshooting or pasting into jsFiddle.}
}
\value{
Creates a temp HTML file and shows it in the viewer.
}
\description{
Allows viewing of hcslim output in RStudio.
}
\examples{
\dontrun{

require(hcslim)

options = list(
 chart = list(type = 'line'),
 title = list(text = 'Some Numbers'),
 yAxis = list(
     title = list(
         text = '123456'
     )
 ),
 series = list(
     list(
       name = '123',
       data = c(1, 2, 3)
     ),
     list(
       name = '456',
       data = c(4, 5, 6)
     )
 )
)

hcslim_view(options)

}
}
