% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter_replace.R
\name{highchart}
\alias{highchart}
\title{Highchart Replacement: highchart}
\usage{
highchart()
}
\value{
List that can be passed to other functions.
}
\description{
Added to make code written with Highcharter work with hcslim.
}
