% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter_replace.R
\name{highchartOutput}
\alias{highchartOutput}
\title{Highchart Replacement: highchartOutput}
\usage{
highchartOutput(id, height = NULL, width = NULL)
}
\arguments{
\item{id}{output ID used in R Shiny.}

\item{height}{desired height of the chart.}

\item{width}{desired width of the chart.}
}
\value{
Output object
}
\description{
Added to make code written with Highcharter work with hcslim.
}
