% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter_replace.R
\name{list_parse}
\alias{list_parse}
\title{Highchart Replacement: list_parse}
\usage{
list_parse(x)
}
\arguments{
\item{x}{Data frame to be converted.}
}
\value{
List representation of data.
}
\description{
Added to make code written with Highcharter work with hcslim.
}
