% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcslim.R
\name{markjs}
\alias{markjs}
\title{Mark JS}
\usage{
markjs(string)
}
\arguments{
\item{string}{}
}
\value{
string with Javascript marked.
}
\description{
Marks Javascript code so hchtml knows how to handle it.
}
