% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options-managers.R
\name{update}
\alias{update}
\title{Update Highcharts Options}
\usage{
update(options, option, ...)
}
\arguments{
\item{options}{Highcharts options list.}

\item{option}{first-level (chart, xAxis, etc.) option to be modified.}

\item{...}{Named second-level options (width, margin, etc.) to be set.}
}
\value{
List that can be passed to other functions.
}
\description{
Preserves current options while adding new ones. This isn't as exact as manually working the list, but may be convenient.
}
