% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcslim.R
\name{usecode}
\alias{usecode}
\title{Import Highcharts Libraries}
\usage{
usecode(...)
}
\arguments{
\item{hcpaths}{Code files to import, minimally defined. For example, to import https://code.highcharts.com/highcharts.js you'd send "highcharts". To import https://code.highcharts.com/modules/accessibility.js you'd send "modules/accessibility".}
}
\description{
Import Highcharts Libraries
}
\examples{
\dontrun{
ui = basicPage(
  usecode(
    'highcharts',
    'modules/exporting',
    'modules/export-data',
    'modules/accessibility'
  ),

  uiOutput('testchart')
)
}
}
