% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcslim.R
\name{hchtml}
\alias{hchtml}
\title{Built HTML for Highchart.}
\usage{
hchtml(
  id,
  options,
  class = c("chart", "mapChart", "stockChart", "ganttChart"),
  loadmapfromurl = NULL,
  printjs = FALSE,
  pretty = printjs
)
}
\arguments{
\item{id}{The HTML DOM id to use for the element and the output key to use. Must be HTML-compatible (no spaces).}

\item{options}{Highcharts options for the chart. Includes data, chart type, etc.}

\item{class}{"chart" in Highcharts.chart(... For Highmaps, it should be "mapChart".}

\item{loadmapfromurl}{If you are using a map, this is the URL of the topo.json file. Example: 'https://code.highcharts.com/mapdata/countries/us/us-ca-all.topo.json'}

\item{printjs}{Print completed JS to console for troubleshooting or pasting into jsFiddle.}

\item{pretty}{Use pretty formatting when converting options to JSON.}
}
\value{
HTML code for calling your chart.
}
\description{
Convers options and places the Highcharts Javascript inside a script tag. Put it inside a renderUI.
}
\examples{
\dontrun{

require(shiny)
require(hcslim)

shinyApp(
 
 ui = basicPage(
   usecode('highcharts'),    
   uiOutput('mychart')
 ),

 server = function(input, output) {

   output$mychart = renderUI({ 
     
     return(hchtml('mychart', list(

       chart = list(type = 'line'),
       title = list(text = 'Some Numbers'),
       yAxis = list(
           title = list(
               text = '123456'
           )
       ),
       series = list(
         list(
           name = '123',
           data = c(1, 2, 3)
         ),
         list(
           name = '456',
           data = c(4, 5, 6)
         )
       )
       
     )))
     
   })

})

}
}
