% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcslim.R
\name{hchtml}
\alias{hchtml}
\title{Built HTML for Highchart.}
\usage{
hchtml(
  id,
  options,
  class = c("chart", "mapChart", "stockChart", "ganttChart"),
  prettyjs = FALSE
)
}
\arguments{
\item{id}{The HTML DOM id to use for the element and the output key to use. Must be HTML-compatible (no spaces).}

\item{options}{Highcharts options for the chart. Includes data, chart type, etc.}

\item{class}{"chart" in Highcharts.chart(... For Highmaps, it should be "mapChart".}

\item{prettyjs}{Use pretty formatting when converting options to JSON.}
}
\value{
HTML code for calling your chart.
}
\description{
Convers options and places the Highcharts Javascript inside a script tag. Put it inside a renderUI.
}
\examples{
\dontrun{
require(shiny)
server = function(input, output) {

  output$testchart = renderUI({ tohc( 'testchart', list(
    chart = list(type = 'line'),
    title = list(text = 'Monthly Average Temperature'),
    subtitle = list(text = 'Source = WorldClimate.com'),
    xAxis = list(
      categories = c(
         'Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 
         'Sep', 'Oct', 'Nov', 'Dec'
      )
    ),
    yAxis = list(
      title= list(
        text = 'Temperature (°C)'
      )
    ),
    plotOptions = list(
      line = list(dataLabels = list(enabled = TRUE), enableMouseTracking = FALSE)
   ),
    series = list(
      list(
        name = 'Tokyo',
        data = c(7.0, 6.9, 9.5, 14.5, 18.4, 21.5, 25.2, 26.5, 23.3, 18.3, 13.9, 9.6)
      ),
      list(
        name = 'London',
        data = c(3.9, 4.2, 5.7, 8.5, 11.9, 15.2, 17.0, 16.6, 14.2, 10.3, 6.6, 4.8)
      )
    )
  ))

})
}
}
}
