% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcslim_view.R
\name{hcslim_view}
\alias{hcslim_view}
\title{View hcslim result in viewer.}
\usage{
hcslim_view(
  options,
  usecode_include = "highcharts",
  class = c("chart", "mapChart", "stockChart", "ganttChart"),
  printjs = FALSE,
  pretty = !printjs
)
}
\arguments{
\item{options}{Highcharts options for the chart. Includes data, chart type, etc.}

\item{usecode_include}{Required Highcharts modules that you'd pass to hcslim::usecode().}

\item{class}{"chart" in Highcharts.chart(... For Highmaps, it should be "mapChart".}

\item{printjs}{Print completed JS to console for troubleshooting or pasting into jsFiddle.}

\item{pretty}{Use pretty formatting when converting options to JSON.}
}
\value{
Creates a temp HTML file and shows it in the browser.
}
\description{
Allows viewing of hcslim output in RStudio.
}
\examples{
\dontrun{

require(hcslim)

options = list(
 chart = list(type = 'line'),
 title = list(text = 'Some Numbers'),
 yAxis = list(
     title = list(
         text = '123456'
     )
 ),
 series = list(
     list(
     name = '123',
     data = c(1, 2, 3)
     ),
     list(
     name = '456',
     data = c(4, 5, 6)
     )
 )
)

hcslim_view(options)

}
}
