% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter_replace.R
\name{renderHighchart}
\alias{renderHighchart}
\title{Highchart Replacement: renderHighchart}
\usage{
renderHighchart(id, options, class = "chart", prettyjs = FALSE)
}
\arguments{
\item{id}{The HTML DOM id to use for the element and the output key to use. Must be HTML-compatible (no spaces).}

\item{options}{Function that builds Highcharts options for the chart.}

\item{class}{"chart" in Highcharts.chart(... For Highmaps, it should be "mapChart".}

\item{prettyjs}{Use pretty formatting when converting options to JSON.}
}
\value{
Render object
}
\description{
Added to make code written with Highcharter work with hcslim.
}
\examples{
\dontrun{
 renderHighchart('mychart', function(){
   # code to build options here (including series data).
 })
}
}
